\name{ActuarialM-package}
\alias{ActuarialM-package}
\docType{package}
\title{
Computation of Actuarial Measures Using Bell G Family
}

\description{
Evaluates the value at risk (VaR) and expected shortfall (ES) of seven well-known probability distributions in connection with the Bell G  family of distributions.
}

\details{
\tabular{ll}{
Package: \tab ActuarialM\cr
Type: \tab Package\cr
Version: \tab 0.1.0 \cr
Date: \tab 2023-05-15\cr
License: \tab GPL-2\cr
}
}

\author{
Muhammad Imran <imranshakoor84@yahoo.com>, M.H. Tahir <mht@iub.edu.pk> and Saima Shakoor <saimashakoor500@gmail.com>.
}

\section{Maintainers }{
Muhammad Imran <imranshakoor84@yahoo.com>
}

\references{
Fayomi, A., Tahir, M. H., Algarni, A., Imran, M., & Jamal, F. (2022). A new useful exponential model with applications to quality control and actuarial data. Computational Intelligence and Neuroscience, 2022. <doi:10.1155/2022/2489998>.

Alsadat, N., Imran, M., Tahir, M. H., Jamal, F., Ahmad, H., & Elgarhy, M. (2023). Compounded Bell-G class of statistical models with applications to COVID-19 and actuarial data. Open Physics, 21(1), 20220242.  <doi:10.1155/2022/2489998>.
}


\keyword{Parameter induction}
\keyword{Compounding}
\keyword{Distribution theory}
\keyword{Generalization of distribution}


